начал верстать AppHeader
