+++ src/components/app-header/app-header.module.css
@@ -1,0 +1,3 @@
+.header{
+  width: 1280px;
+}
