+++ src/styles.module.css
@@ -1,0 +1,5 @@
+.app{
+  width: 1280px;
+  margin: 0 auto;
+  border: 1px solid red;
+}
